FBCM-5157 Filter out `Set-Cookie` headers in `HTTPure.Headers.read`
We must do this in order to avoid runtime type errors when trying to
access `Set-Cookie` request headers. The reason is that node.js
specializes the reading of `Set-Cookie` and makes that one specific
header an `Array String` instead of `String`.

See https://nodejs.org/api/http.html#messageheaders for more details.