Allow multiple SetCookie headers in response (#197)
* Refactor Headers instances and use `show` only for debugging

* Add `HTTPure.MultiHeaders`

This module contains the `MultiHeaders` type which encapsulates the
concept of duplicated headers (or headers with multiple values, like
"Set-Cookie").

* Allow duplicated headers in request and response

We add `multiHeaders` fields to both `HTTPure.Request.Request` and
`HTTPure.Response.Response`, so that headers with multiple values can be
both read from requests and written to responses.

We've chosen this incremental approach instead of changing
`HTTPure.Headers` to avoid a large breaking change.

* Recover the old `show` functionality under `toString`

Some people could have been relying on the old `show` functionality, so
we bring it back under the `toString` name.

* Join headers in both `headers` and `multiHeaders`

Before this commit, headers that existed in both `headers` and
`multiHeaders` were written to the response only with their
`multiHeaders` values. We fix this so that they're joined as if they
were all added to `multiHeaders`.

* Rewrite newly added modules following CitizenNet's style guide

* Fix implementation of `HTTPure.MultiHeaders.read`

When testing the old implementation of this function with an actual
node.js request, the `request.headersDistinct` property did not exist,
even in node.js versions greater than v16.17.0, which, according to
https://nodejs.org/api/http.html#messageheadersdistinct is when the
property was added.

We fix this by using `request.rawHeaders` instead.

* Add the `Examples.MultiHeaders` example

* Filter out `Set-Cookie` headers in `HTTPure.Headers.read`

We must do this in order to avoid runtime type errors when trying to
access `Set-Cookie` request headers. The reason is that node.js
specializes the reading of `Set-Cookie` and makes that one specific
header an `Array String` instead of `String`.

See https://nodejs.org/api/http.html#messageheaders for more details.